🚚 Move UploadWizardModal to /documents/modals
