+++ src/documents/modals/UploadWizard/UploadSteps/ChooseMethodStep.js
