+++ src/documents/modals/UploadWizard/UploadSteps/VersionSummaryStep.js
